# contracts/openapi/ledger-core.public.v1.yaml
openapi: 3.1.0
info:
  title: MAVULA Ledger Core API
  version: 1.0.0
  description: |
    Tenant-scoped accounts, controlled financial adjustments, product configuration, workflows, and read projections.
    Ledger Core is the financial source of truth and enforces idempotency, tenant isolation, and maker-checker controls.
  license:
    name: GNU Affero General Public License v3.0 only
    identifier: AGPL-3.0-only
servers:
  - url: https://ledger.mavula.dev
    description: Ledger Core public endpoint
security: [{ bearerAuth: [] }]
...
paths:
...
  /api/workflows:
    get:
      operationId: listWorkflows
      summary: List workflows
      tags: [Workflows]
      x-mavula-permissions: [finance.read]
      parameters: [{ $ref: '#/components/parameters/WorkflowTrigger' }]
      responses:
        '200': { description: Workflows, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Workflow' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createWorkflow
      summary: Create a workflow
      tags: [Workflows]
      x-mavula-permissions: [configuration.write]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CorrelationId' }]
      requestBody: { $ref: '#/components/requestBodies/CreateWorkflow' }
      responses:
        '201': { description: Workflow created, content: { application/json: { schema: { $ref: '#/components/schemas/Workflow' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/workflows/presets/loan-approval-notification:
    post:
      operationId: createLoanApprovalNotificationWorkflow
      summary: Create the loan approval notification preset
      tags: [Workflows]
      x-mavula-permissions: [configuration.write]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CorrelationId' }]
      responses:
        '201': { description: Preset workflow created, content: { application/json: { schema: { $ref: '#/components/schemas/Workflow' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/workflows/presets/monthly-fee-charge:
    post:
      operationId: createMonthlyFeeChargeWorkflow
      summary: Create the monthly fee charge preset
      tags: [Workflows]
      x-mavula-permissions: [configuration.write]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CorrelationId' }]
      responses:
        '201': { description: Preset workflow created, content: { application/json: { schema: { $ref: '#/components/schemas/Workflow' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/workflows/{workflowId}:
    get:
      operationId: getWorkflow
      summary: Read a workflow
      tags: [Workflows]
      x-mavula-permissions: [finance.read]
      parameters: [{ $ref: '#/components/parameters/WorkflowId' }]
      responses:
        '200': { description: Workflow, content: { application/json: { schema: { $ref: '#/components/schemas/Workflow' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/workflows/{workflowId}/execute:
    post:
      operationId: executeWorkflow
      summary: Execute a workflow
      tags: [Workflows]
      x-mavula-permissions: [finance.write]
      parameters: [{ $ref: '#/components/parameters/WorkflowId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CorrelationId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ExecuteWorkflow' }, example: { context: { account_id: account_001, event: MONTH_END } } } }
      responses:
        '201': { description: Workflow execution completed, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    IdempotencyKey: { in: header, name: Idempotency-Key, required: true, description: Reuse only with the same operation and payload., schema: { type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' } }
    CorrelationId: { in: header, name: X-Correlation-ID, required: false, description: Recommended trace identifier. The server generates one when omitted., schema: { type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' } }
...
    WorkflowId: { in: path, name: workflowId, required: true, schema: { type: string } }
    WorkflowTrigger: { in: query, name: trigger, required: false, description: 'When provided, return only workflows registered for this trigger.', schema: { type: string, pattern: '^[A-Z][A-Z0-9_]{2,100}$' } }
  requestBodies:
...
    CreateWorkflow:
      required: true
      content: { application/json: { schema: { $ref: '#/components/schemas/CreateWorkflow' } } }
  schemas:
    HttpError:
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer }
        message:
          oneOf:
            - { type: string }
            - { type: array, items: { type: string } }
        error: { type: string }
...
    WorkflowStep:
      type: object
      additionalProperties: false
      required: [order, name, action, parameters]
      properties:
        order: { type: integer, minimum: 1 }
        name: { type: string }
        action: { type: string }
        parameters: { type: object, additionalProperties: true }
        condition: { type: string }
    CreateWorkflow:
      type: object
      additionalProperties: false
      required: [name, trigger, steps]
      properties:
        name: { type: string, maxLength: 160 }
        trigger: { type: string, pattern: '^[A-Z][A-Z0-9_]{2,100}$' }
        steps: { type: array, items: { $ref: '#/components/schemas/WorkflowStep' } }
    Workflow:
      allOf:
        - $ref: '#/components/schemas/CreateWorkflow'
        - type: object
          properties:
            id: { type: string }
            tenant_id: { type: string }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    ExecuteWorkflow:
      type: object
      additionalProperties: false
      required: [context]
      properties: { context: { type: object, additionalProperties: true } }
...
  responses:
    BadRequest: { description: Request contract or Idempotency-Key is invalid, content: { application/json: { schema: { $ref: '#/components/schemas/HttpError' } } } }
    Unauthorized: { description: 'Bearer token is missing, expired, invalid, or issued for another audience', content: { application/json: { schema: { $ref: '#/components/schemas/HttpError' } } } }
    Forbidden: { description: 'Permission, maker-checker, institution, or tenant policy denied the operation', content: { application/json: { schema: { $ref: '#/components/schemas/HttpError' } } } }
    NotFound: { description: Tenant-scoped resource was not found, content: { application/json: { schema: { $ref: '#/components/schemas/HttpError' } } } }
    Conflict: { description: State conflict or Idempotency-Key reuse with a different payload, content: { application/json: { schema: { $ref: '#/components/schemas/HttpError' } } } }